+++ src/crash_screen.c
@@ -445,7 +445,7 @@ static s32 crash_screen_print_error(s32 x, s32 y, OSThread* faultedThread) {
         return y;
     }
 
-    u32 badvaddr = (u32)ctx->badvaddr;
+    u32 badvaddr = ctx->badvaddr;
 
     switch (causeIndex) {
         case 1: // TLB modification
@@ -476,7 +476,7 @@ static s32 crash_screen_print_error(s32 x, s32 y, OSThread* faultedThread) {
             y = crash_screen_printf_proportional(x, y, "Bus error at 0x%08lX", badvaddr);
             break;
         case 10: // Reserved instruction
-            y = crash_screen_printf_proportional(x, y, "Invalid instruction at 0x%08lX", (u32)ctx->pc);
+            y = crash_screen_printf_proportional(x, y, "Invalid instruction at 0x%08lX", ctx->pc);
             break;
         case 12: // Arithmetic overflow
             y = crash_screen_printf_proportional(x, y, "Integer overflow");
@@ -580,7 +580,7 @@ void crash_screen_draw(OSThread* faultedThread) {
             y = crash_screen_print_location(x, y, &sym);
             if (isFirstFrame && crashScreenAssertMessage[0] == '\0') {
                 y += 5;
-                y = crash_screen_print_disasm(x, y, (u32)ctx->pc);
+                y = crash_screen_print_disasm(x, y, ctx->pc);
                 y += 5;
                 isFirstFrame = false;
             }
@@ -603,7 +603,7 @@ void crash_screen_draw(OSThread* faultedThread) {
             y = crash_screen_print_location(x, y, &sym);
             if (isFirstFrame && crashScreenAssertMessage[0] == '\0') {
                 y += 5;
-                y = crash_screen_print_disasm(x, y, (u32)ctx->pc);
+                y = crash_screen_print_disasm(x, y, ctx->pc);
                 y += 5;
                 isFirstFrame = false;
             }
